Updating document view in response to Frank's changes. Changed class for font size and moved Documents out of Pages into Agreements.
