+++ frontend/src/components/Agreements/Documents/DocumentView.test.jsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import DocumentView from "./DocumentView";
 
-import store from "../../store";
+import store from "../../../store";
 import { Provider } from "react-redux";
 
 describe("Document View", () => {
